// src/lib/types/diet-plan.ts
export enum Goal {
  GAIN_WEIGHT = "gain weight",
  LOSE_WEIGHT = "lose weight",
  MAINTAIN_WEIGHT = "maintain weight",
  GAIN_MUSCLE = "gain muscle",
  MAINTAIN_MUSCLE = "maintain muscle",
}

export enum Gender {
  MALE = "male",
  FEMALE = "female",
  OTHERS = "others",
}

export enum Preference {
  VEG = "vegetarian",
  VEGAN = "vegan",
  NON_VEG = "non-vegetarian",
  EGGETARIAN = "eggetarian",
}